requirements.txt
Django==4.1.6
django-filter~=22.1